Make 15m and 24h hashrate available after some time
Show 0 for windowed hashrate that is not supposed to be computed yet.

The default behavior was to interpolate hashrate from smaller
time-frame to full size of the window, but this proved to be
confusing as it showed all hashrates (1m, 15m and 24m) the same for
the first minute.

Also, 0 is interpreted by UI as "undefined".